Create README.md
There are 5 words, with maximum of 8 wrong guesses for each. I used dashes in the spaces for the letters because snap seems to have a bug and did not allow me to use underscores. The scratch version (which has less blocks but does the same thing as the snap version) can be found in the MIT repository https://scratch.mit.edu/projects/94412170/#editor
and the snap version is also visible at youtube  https://www.youtube.com/watch?v=Nk4TsUxVzLI