computer code (xml) of hangman project
hangman project in snap (xml) with 5 words with a maximum of 8 wrong guesses allowed. I used dashes because snap seems to have a bug and did not allow me to use underscores. The scratch version is in 
https://scratch.mit.edu/projects/94412170/#player 
and the snap version can be visible at https://www.youtube.com/watch?v=Nk4TsUxVzLI